Support building with th-abstraction-0.6.*
This patch is necessary to build `hobbits` with `template-haskell-2.21.*`,
which is bundled with GHC 9.8. In `template-haskell-2.21.*`, the type of the
type variable binders in `DataD` and `NewtypeD` changes from `[TyVarBndrUnit]`
to `[TyVarBndrSpec]`. Luckily, `th-abstraction-0.6.0.0` backports the
`TyVarBndrSpec` data type, so we can update to `TyVarBndrSpec` in `hobbits`
without needing any CPP.